Wind vane as5048 spi (#1)
* Created a starting package for the rotary sensor.

* Create as5048a node to listen for angle readings

* Restructure Git repository to match master

* Create ROS node for AS5048A

If the file runs as main, it will create a ROS node with publishers and services. The functionality is created to be able to read angle, set zero position, and read diagnostics.

* Changed some documentation and changed how error is handled

Instead of raising exception, a error will be set in the class for the user to handle

* Changed repository structure

Instead of having a src and a vendor directory, all package should be directly below the repository. This is because of confusion of recursive src directories.

* Create unittest setup with modules

__init__.py files marks the packages for python to use when importing modules.

* Split ROS node and class into seperate files

* Created unittest for some functions in the AS5048A class

* Add read bit for NOP address in read and write functions

* Add tests for remaining functionality

* Add functionality to read cordic magnitude

* Add tests for read_cordic_magnitude function

* Add configuration values and a launch file

* Add dependencies for the packages and run setup.py

Uncommented 'catkin_python_setup' to be able to run setup.py from catkin_make

Included some depencies the package depends on, also a python module

* Moved as5048a.py to a package in src

To import custom made modules, such as as5048a.py, I had to move it to a package in src. Because the setup.py searches for packages.

* Create setup.py

Simply to be able to import the as5048a in the scripts and testing purposes.

* Add a mocking feature for testing purposes

* Remove a parameter used for testing

* Replace unittest with ROS unittest

ROS unittest uses unittest, but had to run it from rosunit to be able to save results in a xml format

* Create a .test launch file to run all tests

* Rename from as5048a to as5048

Since the datasheet is for the as5048 and not specifically as5048a

* Add test on services

* Mark executable python scripts

* Fix mistake in test_as5048_service.py

Accidentaly called TestCase instead of the real test case class. Also removed the patching because it doesn't work since the node is run outside the scope.

* Fix error handling for the publisher

* Add comments and docstring

Co-authored-by: MrSurfinbird <sme15002@student.mdh.se>
Co-authored-by: JPDF <pdf16001@student.mdh.se>