Make if_use_heading_as_setpoint a global variable and call it heading_latch. Call calculate_rudder_angle rudder_angle_calculation instead.
Make the if_use_heading_as_setpoint a global variable reduce the inputs in the function is_heading_setpoint, since it both takes it as an input and returns it. Calling it heading_latch is more descriptive when it is a global variable.
The name calculate_rudder_angle is more descriptive of the functionality of the function rudder_angle_calculation.